fiq_fsm: nobble dwc_otg's "error recovery mechanism" with fiq_fsm_enable
If a previous transaction to an endpoint resulted in an error, the dwc_otg
HCD unmasks several interrupts for the following transaction.

This breaks fiq_fsm because it expects a single interrupt to be raised on
transfer completion. Nobble the nominally useless error tracking mechanism
when fiq_fsm is enabled.